refactor: effect path change
