+++ styles/motion/RightSlideEffect.ts
