+++ styles/motion/SpringEffect.ts
@@ -1,10 +1,12 @@
+const defaultEase = [0, 0.71, 0.2, 1.01];
+
 export const SpringEffect = {
   initial: {
     opacity: 0,
     scale: 0.5,
     transition: {
       duration: 0.3,
-      ease: [0, 0.71, 0.2, 1.01],
+      ease: defaultEase,
       scale: {
         type: "spring",
         damping: 7,
@@ -18,7 +20,7 @@ export const SpringEffect = {
     scale: 1,
     transition: {
       duration: 0.3,
-      ease: [0, 0.71, 0.2, 1.01],
+      ease: defaultEase,
       scale: {
         type: "spring",
         damping: 7,
